# public/swagger/user.json
{
    "swagger": "2.0",
    "info": {
        "description": "All APIs require JWT header",
        "version": "1.0.0",
        "title": "Venn user service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "support@getvenn.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:2331",
    "basePath": "/v1",
...
    "schemes": [
        "http"
    ],
    "paths":{
...
        "/user/updateLikes/{userId}":{
            "put":{
                "summary": "Update User Likes",
                "tags": [
                    "User"
                ],
                "description": "",
                "operationId": "UpdateLikes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "type":"number",
                        "description": "Update Likes",
                        "required": true
                    }
                ],
                "responses": {},
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user/getLikes/{userId}":{
            "get":{
                "summary": "Get Likes",
                "tags": [
                    "User"
                ],
                "description": "",
                "operationId": "GetLikes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "type":"number",
                        "description": "Get Likes",
                        "required": true
                    }
                ],
                "responses": {},
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/admin/userDetails/{userId}":{
            "get":{
                "summary": "User Details",
                "tags": [
                    "Admin"
                ],
                "description": "",
                "operationId": "UserDetails",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "userId",
                        "type":"number",
                        "required": true,
                        "description": "User Details"
                    }
                ],
                "responses": {},
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}